backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.3
pandas>=2.2,<3.0
reportlab>=4.2
